requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
boto3==1.24.89
    # via -r requirements.in
botocore==1.27.89
    # via
    #   boto3
    #   s3transfer
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
docker==6.0.0
    # via -r requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.28
    # via -r requirements.in
idna==3.4
    # via requests
jinja2==3.1.2
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via
    #   -r requirements.in
    #   docker
prettytable==3.4.1
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via botocore
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via docker
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
urllib3==1.26.12
    # via
    #   botocore
    #   docker
    #   requests
wcwidth==0.2.5
    # via prettytable
websocket-client==1.4.1
    # via docker
